change int to long long

+++ Assignment 2/CodeGeass.c
@@ -5,31 +5,31 @@
 
 int main() {
 
-  int m,n;
+  long long m,n;
 
-  scanf("%d", &n);
-  scanf("%d", &m);
+  scanf("%lld", &n);
+  scanf("%lld", &m);
 
-  int *knightmares; 
-  int *powers;
+  long long *knightmares; 
+  long long *powers;
 
   knightmares = malloc(n * sizeof(*knightmares));
   powers = malloc(m * sizeof(*powers));
 
-  for(int i = 0; i < n; i++) {
-    scanf("%d", &knightmares[i]);
+  for(long long i = 0; i < n; i++) {
+    scanf("%lld", &knightmares[i]);
   }
 
-  for(int i = 0; i < m; i++) {
-    scanf("%d", &powers[i]);
+  for(long long i = 0; i < m; i++) {
+    scanf("%lld", &powers[i]);
   }
 
-  for(int i = 0; i <= n - m; i++) {
-    for(int j = 0; j < m; j++) {
+  for(long long i = 0; i <= n - m; i++) {
+    for(long long j = 0; j < m; j++) {
       knightmares[j + i] *= powers[j];
     }
   }
 
-  for(int i = 0; i < n; i++)
-    printf("%d ", knightmares[i]);
+  for(long long i = 0; i < n; i++)
+    printf("%lld ", knightmares[i]);
 }
